+++ apps/bus/src/constants/time.ts
@@ -1,0 +1,3 @@
+export const ONE_DAY = 24 * 60 * 60;
+
+export const THIRTY_SEC_IN_MS = 30_000;