+++ apps/bus/src/services/local.ts
@@ -1,11 +1,12 @@
-import type { BusEstimation } from '@f2e/ptx';
+import type { BusEstimationInfo, BusStationInfo } from '@f2e/ptx';
 import { createEntityAdapter, EntityState } from '@reduxjs/toolkit';
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import type { SuccessApiResponse } from 'next-api-handler';
 
-import type { BusEstimationParam } from '@/pages/api/bus-estimation';
+import type { BusEstimationParam } from '@/pages/api/bus/estimation';
+import type { StationQueryParam } from '@/pages/api/bus/nearby';
 
-const busEstimationAdapter = createEntityAdapter<BusEstimation>({
+const busEstimationAdapter = createEntityAdapter<BusEstimationInfo>({
   selectId: (busEstimation) => busEstimation.StopUID,
 });
 
@@ -16,22 +17,28 @@ export const localApi = createApi({
   }),
   endpoints: (builder) => ({
     getBusEstimation: builder.query<
-      EntityState<BusEstimation>,
+      EntityState<BusEstimationInfo>,
       BusEstimationParam
     >({
       query: (params) => ({
-        url: 'bus-estimation',
+        url: 'bus/estimation',
         params,
       }),
-      transformResponse: (res: SuccessApiResponse<BusEstimation[]>) =>
+      transformResponse: (res: SuccessApiResponse<BusEstimationInfo[]>) =>
         busEstimationAdapter.addMany(
           busEstimationAdapter.getInitialState(),
           res.data,
         ),
+    }),
+    getNearByBus: builder.mutation<
+      SuccessApiResponse<BusStationInfo[]>,
+      Record<keyof StationQueryParam, number>
+    >({
+      query: ({ lat, lng }) => `bus/nearby?lat=${lat}&lng=${lng}`,
     }),
   }),
 });
 
 export const busEstimationSelector = busEstimationAdapter.getSelectors();
 
-export const { useGetBusEstimationQuery } = localApi;
+export const { useGetBusEstimationQuery, useGetNearByBusMutation } = localApi;